+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 Follows the format convention here: [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2020-07-22
+
+[This Release](http://s3.amazonaws.com/socotra.wb/liquid-playground/v0.3.0/index.html)
+|
+[Latest](http://s3.amazonaws.com/socotra.wb/liquid-playground/index.html)
+
+### Added
+
+- Error messages on Liquid rendering errors
+- Usage Tracking via Segment and Google Analytics
+
 ## [0.2.0] - 2020-07-17
 
 [This Release](http://s3.amazonaws.com/socotra.wb/liquid-playground/v0.2.0/index.html)